+++ CHANGELOG.md
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - SimpleDecal の IslandCulling を実験的機能に変更 `#229`
 - SimpleDecal の Far Culling の基準を変更 `#242`
 - PropertyName の保存形式を調整し、データが勝手に変わらないように変更 `#228`
+- Texture のブレンドを即時実行するように変更し、VRAM + RAM 容量が Decalなどのコンポーネントの最大数にならないように変更 `#188`
 
 ### Removed
 
@@ -45,6 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - 一部の色合成が一般的なソフトと大きく異なっている問題を修正 `#237`
 - AtlasTexture の FineTuning の Resize の品質が低い問題を修正 `#96`
+- Decalなどのコンポーネントがオリジナルのテクスチャーを取得する際、不必要に別のインスタンスを生成していた問題を修正 `#249`
 
 ### Deprecated